Fixed errors and implemented delete and update item in cart

+++ src/main/java/store/repository/CartItemRepository.java
@@ -3,5 +3,5 @@ package store.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import store.model.CartItem;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 }
