+++ src/main/java/store/repository/ShoppingCartRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import store.model.ShoppingCart;
 
-public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
+public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
     Optional<ShoppingCart> findByUserId(Long id);
 }
